+++ packages/yoastseo/src/languageProcessing/languages/sv/Researcher.js
@@ -1,18 +1,16 @@
-import matchKeywordInSubheadings from "../../researches/base/matchKeywordInSubheadings.js";
-import getProminentWordsForInsights from "../../researches/base/getProminentWordsForInsights";
-import getProminentWordsForInternalLinking from "../../researches/base/getProminentWordsForInternalLinking";
-import getWordForms from "../../researches/_todo/getWordForms";
-import findKeywordInPageTitle from "../../researches/base/findKeywordInPageTitle";
-import { keyphraseDistributionResearcher as keyphraseDistribution } from "../../researches/base/keyphraseDistribution";
-
 import AbstractResearcher from "../../AbstractResearcher";
 
-import getLinkStatistics from "./researches/getLinkStatistics";
-import passiveVoice from "./researches/getPassiveVoice.js";
-import findTransitionWords from "./researches/findTransitionWords";
-import functionWordsInKeyphrase from "./researches/functionWordsInKeyphrase";
-import getSentenceBeginnings from "./researches/getSentenceBeginnings.js";
-import keyphraseLength from "./researches/keyphraseLength";
+// All config
+import firstWordExceptions from "./config/firstWordExceptions";
+import functionWords from "./config/functionWords";
+import transitionWords from "./config/transitionWords";
+import twoPartTransitionWords from "./config/twoPartTransitionWords";
+
+// All helpers
+
+import getStemmer from "../../helpers/morphology/baseStemmer";
+import isPassiveSentence from "./helpers/isPassiveSentence";
+
 
 /**
  * The researches contains all the researches
@@ -26,19 +24,23 @@ export default class Researcher extends AbstractResearcher {
 	constructor( paper ) {
 		super( paper );
 
-		Object.assign( this.defaultResearches, {
-			findTransitionWords: findTransitionWords,
-			passiveVoice: passiveVoice,
-			getSentenceBeginnings: getSentenceBeginnings,
-			functionWordsInKeyphrase: functionWordsInKeyphrase,
-			matchKeywordInSubheadings: matchKeywordInSubheadings,
-			getLinkStatistics: getLinkStatistics,
-			keyphraseDistribution: keyphraseDistribution,
-			findKeywordInPageTitle: findKeywordInPageTitle,
-			morphology: getWordForms,
-			prominentWordsForInsights: getProminentWordsForInsights,
-			prominentWordsForInternalLinking: getProminentWordsForInternalLinking,
-			keyphraseLength: keyphraseLength,
+		// Delete the researches that are not available in Swedish.
+		delete this.defaultResearches.getFleschReadingScore;
+		delete this.defaultResearches.stopWordsInKeyword;
+		delete this.defaultResearches.stopWordsInUrl;
+
+		Object.assign( this.config, {
+			language: "sv",
+			isPeriphrastic: false,
+			firstWordExceptions,
+			functionWords,
+			transitionWords,
+			twoPartTransitionWords,
+		} );
+
+		Object.assign( this.helpers, {
+			getStemmer,
+			isPassiveSentence,
 		} );
 	}
 }
