+++ packages/yoastseo/src/languageProcessing/languages/sv/config/internal/participles.js
