+++ packages/yoastseo/src/languageProcessing/languages/sv/helpers/isPassiveSentence.js
@@ -1,5 +1,5 @@
 import areWordsInSentence from "../../../helpers/word/areWordsInSentence";
-import getPassiveVerbs from "../config/passiveVoice/participles";
+import getPassiveVerbs from "../config/internal/participles";
 
 /**
  * Checks the passed sentence to see if it contains passive verbs.
