--- packages/yoastseo/src/languageProcessing/languages/sv/researches/matchKeywordInSubheadings.js
@@ -1,10 +1,0 @@
-import matchKeywordInSubheadings from "../../../researches/base/matchKeywordInSubheadings";
-
-/**
- * Checks if there are any h2 or h3 subheadings in the text and if they have the keyphrase or synonyms in them.
- *
- * @inheritDoc matchKeywordInSubheadings
- */
-export default function( paper, researcher ) {
-	return matchKeywordInSubheadings( paper, researcher, true );
-}
